@@ -4,11 +4,16 @@
 Convex Contract
 
 """
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from convex_api.api import API
 
 import re
-from typing import Union
+
 from convex_api.account import Account
-from convex_api.api import API
 
 
 class Contract:
@@ -25,9 +30,9 @@ class Contract:
 
     def load(
         self,
-        name: Union[str, None] = None,
-        address: Union[Account, int, str, None] = None,
-        owner_address: Union[Account, int, str, None] = None
+        name: str | None = None,
+        address: Account | int | str | None = None,
+        owner_address: Account | int | str | None = None
     ):
         """
 
@@ -39,7 +44,7 @@ class Contract:
         :param str, int, Account address: Address of the contract, if the name is not known,
         then you can provide the actual address of the contract.
 
-        :param str, int, Account owner_address: If the contract is registered the owner address of the resgistration.
+        :param str, int, Account owner_address: If the contract is registered the owner address of the registration.
 
         :returns int The address of the resolved contract
 
@@ -62,10 +67,10 @@ class Contract:
     def deploy(
         self,
         account: Account,
-        text: Union[str, None] = None,
-        filename: Union[str, None] = None,
-        name: Union[str, None] = None,
-        owner_account: Union[Account, None] = None
+        text: str | None = None,
+        filename: str | None = None,
+        name: str | None = None,
+        owner_account: Account | None = None
     ):
         """
 
@@ -79,7 +84,7 @@ class Contract:
 
         :param str name: Name of the contract to register
 
-        :param Account onwer_account: Optional owner account of the registration.
+        :param Account owner_account: Optional owner account of the registration.
         If not provided then the Account will be used.
 
         :returns Address of the new contract
@@ -141,7 +146,7 @@ class Contract:
             raise ValueError(f'No contract address found for {self._name}')
         return self._convex.send(f'(call #{self._address} {transaction})', account)
 
-    def query(self, transaction: str, account_address: Union[Account, int, str, None] = None):
+    def query(self, transaction: str, account_address: Account | int | str | None = None):
         """
 
         Sends a query to the contract.
