@@ -5,11 +5,8 @@
 """
 
 
-from typing import Union
-
-
 class ConvexBaseError(Exception):
-    def __init__(self, source: str, code: Union[int, str], text: str):
+    def __init__(self, source: str, code: int | str, text: str):
         self.source = source
         self.code = code
         self.text = text
