@@ -5,10 +5,7 @@
 """
 
 from argparse import Namespace
-from typing import (
-    Literal,
-    Union
-)
+from typing import Literal
 
 from convex_api.tool.command.argparse_typing import (
     BaseArgs,
@@ -23,12 +20,12 @@ class AccountFundArgs(BaseArgs):
     command: Literal['account']
     account_command: Literal['fund']
     amount: int
-    name_address: Union[str, int]
+    name_address: str | int
 
 
 class AccountFundCommand(CommandBase):
 
-    def __init__(self, sub_parser: Union[SubParsersAction, None] = None):
+    def __init__(self, sub_parser: SubParsersAction | None = None):
         self._command_list = []
         super().__init__('fund', sub_parser)
 
