@@ -4,13 +4,10 @@ from argparse import (
 )
 from typing import (
     Any,
-    Dict,
-    Protocol,
-    Union
+    Protocol
 )
 
 from pydantic import (
-    AnyHttpUrl,
     BaseModel,
     Field
 )
@@ -18,21 +15,21 @@ from pydantic import (
 
 class SubParsersAction(Protocol):
     @property
-    def choices(self) -> Dict[str, ArgumentParser]:
+    def choices(self) -> dict[str, ArgumentParser]:
         ...
 
     def add_parser(self, name: str, **kwargs: Any) -> ArgumentParser:
         ...
 
-    def __call__(self, parser: ArgumentParser, namespace: Namespace, values: Any, option_string: Union[str, None] = None) -> None:
+    def __call__(self, parser: ArgumentParser, namespace: Namespace, values: Any, option_string: str | None = None) -> None:
         ...
 
 
 class BaseArgs(BaseModel):
-    keyfile: Union[None, str]
-    keytext: Union[None, str]
-    password: Union[None, str]
-    keywords: Union[None, str]
+    keyfile: None | str = None
+    keytext: None | str = None
+    password: None | str = None
+    keywords: None | str = None
     debug: bool = False
     output_json: bool = Field(alias='json', default=False)
-    url: AnyHttpUrl
+    url: str
