@@ -3,15 +3,16 @@
 
 """
 import json
+
 import requests
-
-from cryptography.hazmat.primitives.asymmetric.ed25519 import Ed25519PrivateKey
 from cryptography.hazmat.primitives import serialization
-from convex_api.key_pair import KeyPair
+from cryptography.hazmat.primitives.asymmetric.ed25519 import Ed25519PrivateKey
 
+from convex_api.key_pair import KeyPair
 
 PRIVATE_TEST_KEY = 0x973f69bcd654b264759170724e1e30ccd2e75fc46b7993fd24ce755f0a8c24d0
 PUBLIC_ADDRESS = '5288fec4153b702430771dfac8aed0b21cafca4344dae0d47b97f0bf532b3306'
+
 
 def get_test_account():
     private_key = Ed25519PrivateKey.from_private_bytes(KeyPair.to_bytes(PRIVATE_TEST_KEY))
@@ -22,8 +23,9 @@ def get_test_account():
         format=serialization.PublicFormat.Raw
     )
     public_address = KeyPair.remove_0x_prefix(KeyPair.to_hex(public_key_bytes))
-    assert(public_address == PUBLIC_ADDRESS)
+    assert public_address == PUBLIC_ADDRESS
     return private_key, public_address
+
 
 def generate_test_keys():
     private_key = Ed25519PrivateKey.generate()
@@ -43,9 +45,10 @@ def create_account(convex_url: str, public_address: str):
     url = f'{convex_url}/api/v1/createAccount'
     print('create_account send to ', url, 'data:', account_data)
     response = requests.post(url, data=json.dumps(account_data))
-    assert(response.status_code == 200)
+    assert response.status_code == 200
     result = response.json()
     return result['address']
+
 
 def request_funds(convex_url: str, address: str):
     faucet_data = {
@@ -57,7 +60,7 @@ def request_funds(convex_url: str, address: str):
     response = requests.post(url, data=json.dumps(faucet_data))
     if response.status_code != 200:
         print('error', response.text)
-        assert(response.status_code == 200)
+        assert response.status_code == 200
     print('faucet response', response.json())
 
 
@@ -78,11 +81,11 @@ def test_submit_transaction(convex_url: str):
     response = requests.post(url, data=json.dumps(prepare_data))
     if response.status_code != 200:
         print('prepare error', response.text)
-        assert(response.status_code == 200)
+        assert response.status_code == 200
 
     result = response.json()
 
-    #submit
+    # submit
     print(result)
     hash_data = KeyPair.hex_to_bytes(result['hash'])
     signed_hash_bytes = private_key.sign(hash_data)
@@ -99,9 +102,8 @@ def test_submit_transaction(convex_url: str):
     response = requests.post(url, data=json.dumps(submit_data))
     if response.status_code != 200:
         print('submit error', response.text, response.status_code)
-        assert(response.status_code == 200)
+        assert response.status_code == 200
     print(response.json())
-
 
 
 def test_query_lisp(convex_url: str):
@@ -118,10 +120,10 @@ def test_query_lisp(convex_url: str):
     response = requests.post(url, data=json.dumps(query_data))
     if response.status_code != 200:
         print('query error', response.text)
-        assert(response.status_code == 200)
+        assert response.status_code == 200
     result = response.json()
-    assert(result)
-    assert(result['value'] > 0)
+    assert result
+    assert result['value'] > 0
 
 
 def DISABLED_test_query_scrypt(convex_url: str):
@@ -138,9 +140,8 @@ def DISABLED_test_query_scrypt(convex_url: str):
     response = requests.post(url, data=json.dumps(query_data))
     if response.status_code != 200:
         print('query error', response.text)
-        assert(response.status_code == 200)
+        assert response.status_code == 200
     result = response.json()
-    assert(result)
+    assert result
     print(result)
-    assert(result['value'] > 0)
-
+    assert result['value'] > 0
