@@ -5,8 +5,9 @@
 
 """
 
-import pytest
 import secrets
+
+import pytest
 
 from convex_api.account import Account
 from convex_api.api import API
@@ -50,7 +51,7 @@ def test_convex_recursion(convex: API, test_account: Account):
 """
         convex.topup_account(test_account)
         result = convex.send(contract, test_account)
-        assert(result is not None)
+        assert result is not None
         address_list.append(Account.to_address(result.value))
     for index in range(0, chain_length):
         next_index = index + 1
@@ -64,10 +65,11 @@ def test_convex_recursion(convex: API, test_account: Account):
                 result = convex.send(f'(call chain-{index} (set {test_number}))', test_account)
         else:
             result = convex.send(f'(call chain-0 (set {test_number}))', test_account)
-            assert(result is not None)
-            assert(result.value == test_number)
+            assert result is not None
+            assert result.value == test_number
     with pytest.raises(ConvexAPIError, match='DEPTH'):
         convex.query('(call chain-0 (get))', test_account)
+
 
 def test_schedule_transfer(convex: API, test_account: Account, other_account: Account):
     # you can send coins to an actor , if it exports the receive-coin function
@@ -107,7 +109,7 @@ def test_schedule_transfer(convex: API, test_account: Account, other_account: Ac
     convex.topup_account(test_account)
     convex.topup_account(other_account, 8000000)
     result = convex.send(contract, test_account)
-    assert(result is not None)
+    assert result is not None
     contract_address = Account.to_address(result.value)
     convex.transfer(contract_address, 800000, other_account)
     convex.topup_account(test_account)
